refactor(popup): split manifest from component registry
Related to #434
