+++ apps/desktop/src/contracts/index.ts
@@ -4,6 +4,8 @@ export { AiError, AiErrorCode } from './agentErrors';
 export type * from './attachments';
 export type * from './builtInToolPresentation';
 export type * from './json';
+export type * from './popup';
+export type * from './popupManifest';
 export type * from './providerConfig';
 export type * from './sessionStatus';
 export type * from './tooling';
